+++ kibot/fil_var_rename_kicost.py
@@ -20,7 +20,8 @@ logger = log.get_logger(__name__)
 class Var_Rename_KiCost(BaseFilter):  # noqa: F821
     """ Var_Rename_KiCost
         This filter implements the kicost.VARIANT:FIELD=VALUE renamer to get FIELD=VALUE when VARIANT is in use.
-        It applies the KiCost concept of variants (a regex to match the VARIANT) """
+        It applies the KiCost concept of variants (a regex to match the VARIANT).
+        The internal `_var_rename_kicost` filter emulates the KiCost behavior """
     def __init__(self):
         super().__init__()
         self._is_transform = True
@@ -53,7 +54,7 @@ class Var_Rename_KiCost(BaseFilter):  # noqa: F821
                 variant = GS.variant[0]
             else:
                 variant = '('+'|'.join(GS.variant)+')'
-        var = re.compile(variant, re.IGNORECASE)
+        var_re = re.compile(variant, re.IGNORECASE)
         for name, value in comp.get_user_fields():
             name = name.strip().lower()
             # Remove the prefix
@@ -69,13 +70,13 @@ class Var_Rename_KiCost(BaseFilter):  # noqa: F821
                 # Successfully separated
                 f_variant = res[0].lower()
                 f_field = res[1].lower()
-                if var.match(f_variant):
+                if var_re.match(f_variant):
                     # Variant matched
                     if GS.debug_level > 2:
                         logger.debug('ref: {} {}: {} -> {}'.
                                      format(comp.ref, f_field, comp.get_field_value(f_field), value))
                     comp.set_field(f_field, value)
-            elif self.variant_to_value and var.match(name):
+            elif self.variant_to_value and var_re.match(name):
                 # The field matches the variant and the user wants to change the value
                 if GS.debug_level > 2:
                     logger.debug('ref: {} value: {} -> {}'.format(comp.ref, comp.value, value))