+++ kibot/misc.py
@@ -79,8 +79,10 @@ KICAD_VERSION_5_99 = 5099000
 # Internal filter names
 IFILT_MECHANICAL = '_mechanical'
 IFILT_VAR_RENAME = '_var_rename'
+IFILT_VAR_RENAME_KICOST = '_var_rename_kicost'
 IFILT_ROT_FOOTPRINT = '_rot_footprint'
 IFILT_KICOST_RENAME = '_kicost_rename'
+IFILT_KICOST_DNP = '_kicost_dnp'
 # KiCad 5 GUI values for the attribute
 UI_THT = 0       # 1 for KiCad 6
 UI_SMD = 1       # 2 for KiCad 6